libxapi : stage error tables with constructor functions
Eliminates the need to make a fatal call to xapi_errtab_register for every error
table, while imposing a limit on the number of error tables that may be used in
this way, since it uses static storage. Registration of tables at runtime is
still supported by xapi_errtab_register, and uses a dynamic allocation.

Signed-off-by: Todd Freed <todd.freed@gmail.com>